# Fix performance of System.Security. Cryptography.Tests