--- a/src/libraries/Common/tests/System/Security/Cryptography/AlgorithmImplementations/RSA/ImportExport.cs
+++ b/src/libraries/Common/tests/System/Security/Cryptography/AlgorithmImplementations/RSA/ImportExport.cs
@@ -10,7 +10,8 @@ namespace System.Security.Cryptography.Rsa.Tests
     [SkipOnPlatform(TestPlatforms.Browser, "Not supported on Browser")]
     public partial class ImportExport
     {
-        public static bool Supports16384 { get; } = TestRsa16384();
+        private static readonly Lazy<bool> s_supports16384 = new Lazy<bool>(TestRsa16384);
+        public static bool Supports16384 => s_supports16384.Value;
 
         [Fact]
         public static void ExportAutoKey()